requirements.txt
Appium-Python-Client>=3.1.0
selenium>=4.0.0
flask
requests
python-dotenv
openai
fastapi
uvicorn
httpx
anthropic
redis
apache-airflow-providers-redis
smbprotocol
ultralytics>=8.0.0
